Fix issues with url tag and email subject
The url tag changed in Django 1.5. I fixed the code so that it should
work in both Django 1.4 and 1.5. When we drop support for 1.4, we can
ditch all the

    {% load url from future %}

stuff.

The email subject for deferred awards is a template which is fine but
when we render the template, it's likely to end up with newlines. So I
added a call to strip() which will remove whitespace at the beginning
and end. It fixes the test, but there might be other things we could do
better here. For example, maybe use gettext instead of putting it in a
template?